Fixed: Rename parser helper class
