+++ source/js/plugins/metadata.js
@@ -20,7 +20,7 @@
  * Toolbar plugins must have a toolbarIcon and toolbarSide attribute, with toolbarSide being either 'left' or 'right'
  **/
 
-import parseLabelValue from '../utils/label-value-parser';
+import parseLabelValue from '../utils/parse-label-value';
 
 export default class MetadataPlugin
 {
