+++ source/js/utils/parse-label-value.js
